move Tooltip into it's own CS file
